requirements.txt
google-genai
python-dotenv
wikipedia
openmeteo-requests
requests-cache
retry-requests
